Initial Draft of Delegation Removal
