+++ genesis/genesis.go
@@ -406,8 +406,6 @@ func FromConfig(config *Config) ([]byte, ids.ID, error) {
 			}
 		}
 
-		delegationFee := json.Uint32(staker.DelegationFee)
-
 		platformvmArgs.Validators = append(platformvmArgs.Validators,
 			platformvm.APIPrimaryValidator{
 				APIStaker: platformvm.APIStaker{
@@ -419,8 +417,7 @@ func FromConfig(config *Config) ([]byte, ids.ID, error) {
 					Threshold: 1,
 					Addresses: []string{destAddrStr},
 				},
-				Staked:             utxos,
-				ExactDelegationFee: &delegationFee,
+				Staked: utxos,
 			},
 		)
 	}
