+++ genesis/unparsed_config.go
@@ -68,13 +68,10 @@ func (ua UnparsedAllocation) Parse() (Allocation, error) {
 type UnparsedStaker struct {
 	NodeID        string `json:"nodeID"`
 	RewardAddress string `json:"rewardAddress"`
-	DelegationFee uint32 `json:"delegationFee"`
 }
 
 func (us UnparsedStaker) Parse() (Staker, error) {
-	s := Staker{
-		DelegationFee: us.DelegationFee,
-	}
+	s := Staker{}
 
 	nodeID, err := ids.ShortFromPrefixedString(us.NodeID, constants.NodeIDPrefix)
 	if err != nil {
