+++ vms/platformvm/cache_current_validator.go
@@ -10,15 +10,10 @@ type currentValidator interface {
 
 	AddValidatorTx() *UnsignedAddValidatorTx
 
-	// Weight of delegations to this validator. Doesn't include the stake
-	// provided by this validator.
-	DelegatorWeight() uint64
-
 	PotentialReward() uint64
 }
 
 type currentValidatorImpl struct {
-	// delegators are sorted in order of removal.
 	validatorImpl
 
 	addValidatorTx  *UnsignedAddValidatorTx
