+++ vms/platformvm/cache_internal_state.go
@@ -909,14 +909,6 @@ func (st *internalStateImpl) writeCurrentStakers() error {
 			subnetID = constants.PrimaryNetworkID
 			nodeID = tx.Validator.NodeID
 			weight = tx.Validator.Wght
-		case *UnsignedAddDelegatorTx:
-			if err := database.PutUInt64(st.currentDelegatorList, txID[:], potentialReward); err != nil {
-				return err
-			}
-
-			subnetID = constants.PrimaryNetworkID
-			nodeID = tx.Validator.NodeID
-			weight = tx.Validator.Wght
 		case *UnsignedAddSubnetValidatorTx:
 			if err := st.currentSubnetValidatorList.Put(txID[:], nil); err != nil {
 				return err
@@ -962,12 +954,6 @@ func (st *internalStateImpl) writeCurrentStakers() error {
 
 			delete(st.uptimes, tx.Validator.NodeID)
 			delete(st.updatedUptimes, tx.Validator.NodeID)
-
-			subnetID = constants.PrimaryNetworkID
-			nodeID = tx.Validator.NodeID
-			weight = tx.Validator.Wght
-		case *UnsignedAddDelegatorTx:
-			db = st.currentDelegatorList
 
 			subnetID = constants.PrimaryNetworkID
 			nodeID = tx.Validator.NodeID
@@ -1072,8 +1058,6 @@ func (st *internalStateImpl) writePendingStakers() error {
 		switch tx.UnsignedTx.(type) {
 		case *UnsignedAddValidatorTx:
 			db = st.pendingValidatorList
-		case *UnsignedAddDelegatorTx:
-			db = st.pendingDelegatorList
 		case *UnsignedAddSubnetValidatorTx:
 			db = st.pendingSubnetValidatorList
 		default:
@@ -1092,8 +1076,6 @@ func (st *internalStateImpl) writePendingStakers() error {
 		switch tx.UnsignedTx.(type) {
 		case *UnsignedAddValidatorTx:
 			db = st.pendingValidatorList
-		case *UnsignedAddDelegatorTx:
-			db = st.pendingDelegatorList
 		case *UnsignedAddSubnetValidatorTx:
 			db = st.pendingSubnetValidatorList
 		default:
@@ -1348,47 +1330,6 @@ func (st *internalStateImpl) loadCurrentValidators() error {
 	if err := validatorIt.Error(); err != nil {
 		return err
 	}
-
-	delegatorIt := st.currentDelegatorList.NewIterator()
-	defer delegatorIt.Release()
-	for delegatorIt.Next() {
-		txIDBytes := delegatorIt.Key()
-		txID, err := ids.ToID(txIDBytes)
-		if err != nil {
-			return err
-		}
-		tx, _, err := st.GetTx(txID)
-		if err != nil {
-			return err
-		}
-
-		potentialRewardBytes := delegatorIt.Value()
-		potentialReward, err := database.ParseUInt64(potentialRewardBytes)
-		if err != nil {
-			return err
-		}
-
-		addDelegatorTx, ok := tx.UnsignedTx.(*UnsignedAddDelegatorTx)
-		if !ok {
-			return errWrongTxType
-		}
-
-		cs.validators = append(cs.validators, tx)
-		vdr, exists := cs.validatorsByNodeID[addDelegatorTx.Validator.NodeID]
-		if !exists {
-			return errDelegatorSubset
-		}
-		vdr.delegatorWeight += addDelegatorTx.Validator.Wght
-		vdr.delegators = append(vdr.delegators, addDelegatorTx)
-		cs.validatorsByTxID[txID] = &validatorReward{
-			addStakerTx:     tx,
-			potentialReward: potentialReward,
-		}
-	}
-	if err := delegatorIt.Error(); err != nil {
-		return err
-	}
-
 	subnetValidatorIt := st.currentSubnetValidatorList.NewIterator()
 	defer subnetValidatorIt.Release()
 	for subnetValidatorIt.Next() {
@@ -1421,9 +1362,6 @@ func (st *internalStateImpl) loadCurrentValidators() error {
 		return err
 	}
 
-	for _, vdr := range cs.validatorsByNodeID {
-		sortDelegatorsByRemoval(vdr.delegators)
-	}
 	sortValidatorsByRemoval(cs.validators)
 	cs.setNextStaker()
 
@@ -1475,20 +1413,7 @@ func (st *internalStateImpl) loadPendingValidators() error {
 			return err
 		}
 
-		addDelegatorTx, ok := tx.UnsignedTx.(*UnsignedAddDelegatorTx)
-		if !ok {
-			return errWrongTxType
-		}
-
 		ps.validators = append(ps.validators, tx)
-		if vdr, exists := ps.validatorExtrasByNodeID[addDelegatorTx.Validator.NodeID]; exists {
-			vdr.delegators = append(vdr.delegators, addDelegatorTx)
-		} else {
-			ps.validatorExtrasByNodeID[addDelegatorTx.Validator.NodeID] = &validatorImpl{
-				delegators: []*UnsignedAddDelegatorTx{addDelegatorTx},
-				subnets:    make(map[ids.ID]*UnsignedAddSubnetValidatorTx),
-			}
-		}
 	}
 	if err := delegatorIt.Error(); err != nil {
 		return err
@@ -1525,10 +1450,6 @@ func (st *internalStateImpl) loadPendingValidators() error {
 	}
 	if err := subnetValidatorIt.Error(); err != nil {
 		return err
-	}
-
-	for _, vdr := range ps.validatorExtrasByNodeID {
-		sortDelegatorsByAddition(vdr.delegators)
 	}
 	sortValidatorsByAddition(ps.validators)
