+++ vms/platformvm/cache_validator.go
@@ -21,19 +21,12 @@ import (
 var _ validator = &validatorImpl{}
 
 type validator interface {
-	Delegators() []*UnsignedAddDelegatorTx
 	SubnetValidators() map[ids.ID]*UnsignedAddSubnetValidatorTx
 }
 
 type validatorImpl struct {
-	// sorted in order of next operation, either addition or removal.
-	delegators []*UnsignedAddDelegatorTx
 	// maps subnetID to tx
 	subnets map[ids.ID]*UnsignedAddSubnetValidatorTx
-}
-
-func (v *validatorImpl) Delegators() []*UnsignedAddDelegatorTx {
-	return v.delegators
 }
 
 func (v *validatorImpl) SubnetValidators() map[ids.ID]*UnsignedAddSubnetValidatorTx {
