+++ vms/platformvm/codec.go
@@ -62,7 +62,6 @@ func init() {
 
 			c.RegisterType(&UnsignedAddValidatorTx{}),
 			c.RegisterType(&UnsignedAddSubnetValidatorTx{}),
-			c.RegisterType(&UnsignedAddDelegatorTx{}),
 
 			c.RegisterType(&UnsignedCreateChainTx{}),
 			c.RegisterType(&UnsignedCreateSubnetTx{}),
