+++ vms/platformvm/factory.go
@@ -63,12 +63,6 @@ type Factory struct {
 	// The maximum amount of tokens that can be bonded on a validator
 	MaxValidatorStake uint64
 
-	// Minimum stake, in nAVAX, that can be delegated on the primary network
-	MinDelegatorStake uint64
-
-	// Minimum fee that can be charged for delegation
-	MinDelegationFee uint32
-
 	// UptimePercentage is the minimum uptime required to be rewarded for staking
 	UptimePercentage float64
