+++ vms/platformvm/service_test.go
@@ -103,7 +103,7 @@ func defaultAddress(t *testing.T, service *Service) {
 }
 
 func TestAddValidator(t *testing.T) {
-	expectedJSONString := `{"username":"","password":"","from":null,"changeAddr":"","txID":"11111111111111111111111111111111LpoYY","startTime":"0","endTime":"0","nodeID":"","rewardAddress":"","delegationFeeRate":"0.0000"}`
+	expectedJSONString := `{"username":"","password":"","from":null,"changeAddr":"","txID":"11111111111111111111111111111111LpoYY","startTime":"0","endTime":"0","nodeID":"","rewardAddress":""}`
 	args := AddValidatorArgs{}
 	bytes, err := json.Marshal(&args)
 	if err != nil {
@@ -347,7 +347,6 @@ func TestGetTx(t *testing.T) {
 					uint64(service.vm.clock.Time().Add(syncBound).Add(defaultMinStakingDuration).Unix()),
 					ids.GenerateTestShortID(),
 					ids.GenerateTestShortID(),
-					0,
 					[]*crypto.PrivateKeySECP256K1R{keys[0]},
 					keys[0].PublicKey().Address(), // change addr
 				)
@@ -470,288 +469,7 @@ func TestGetBalance(t *testing.T) {
 }
 
 // Test method GetStake
-func TestGetStake(t *testing.T) {
-	assert := assert.New(t)
-	service := defaultService(t)
-	defaultAddress(t, service)
-	service.vm.ctx.Lock.Lock()
-	defer func() {
-		err := service.vm.Shutdown()
-		assert.NoError(err)
-		service.vm.ctx.Lock.Unlock()
-	}()
-
-	// Ensure GetStake is correct for each of the genesis validators
-	genesis, _ := defaultGenesis()
-	addrsStrs := []string{}
-	for i, validator := range genesis.Validators {
-		addr := fmt.Sprintf("P-%s", validator.RewardOwner.Addresses[0])
-		addrsStrs = append(addrsStrs, addr)
-		args := GetStakeArgs{
-			api.JSONAddresses{
-				Addresses: []string{addr},
-			},
-			formatting.Hex,
-		}
-		response := GetStakeReply{}
-		err := service.GetStake(nil, &args, &response)
-		assert.NoError(err)
-		assert.EqualValues(uint64(defaultWeight), uint64(response.Staked))
-		assert.Len(response.Outputs, 1)
-		// Unmarshal into an output
-		outputBytes, err := formatting.Decode(args.Encoding, response.Outputs[0])
-		assert.NoError(err)
-		var output avax.TransferableOutput
-		_, err = Codec.Unmarshal(outputBytes, &output)
-		assert.NoError(err)
-		out, ok := output.Out.(*secp256k1fx.TransferOutput)
-		assert.True(ok)
-		assert.EqualValues(out.Amount(), defaultWeight)
-		assert.EqualValues(out.Threshold, 1)
-		assert.Len(out.Addrs, 1)
-		assert.Equal(keys[i].PublicKey().Address(), out.Addrs[0])
-		assert.EqualValues(out.Locktime, 0)
-	}
-
-	// Make sure this works for multiple addresses
-	args := GetStakeArgs{
-		api.JSONAddresses{
-			Addresses: addrsStrs,
-		},
-		formatting.Hex,
-	}
-	response := GetStakeReply{}
-	err := service.GetStake(nil, &args, &response)
-	assert.NoError(err)
-	assert.EqualValues(len(genesis.Validators)*defaultWeight, response.Staked)
-	assert.Len(response.Outputs, len(genesis.Validators))
-	for _, outputStr := range response.Outputs {
-		outputBytes, err := formatting.Decode(args.Encoding, outputStr)
-		assert.NoError(err)
-		var output avax.TransferableOutput
-		_, err = Codec.Unmarshal(outputBytes, &output)
-		assert.NoError(err)
-		out, ok := output.Out.(*secp256k1fx.TransferOutput)
-		assert.True(ok)
-		assert.EqualValues(defaultWeight, out.Amount())
-		assert.EqualValues(out.Threshold, 1)
-		assert.EqualValues(out.Locktime, 0)
-		assert.Len(out.Addrs, 1)
-	}
-
-	oldStake := uint64(defaultWeight)
-
-	// Add a delegator
-	stakeAmt := service.vm.MinDelegatorStake + 12345
-	delegatorNodeID := keys[0].PublicKey().Address()
-	delegatorEndTime := uint64(defaultGenesisTime.Add(defaultMinStakingDuration).Unix())
-	tx, err := service.vm.newAddDelegatorTx(
-		stakeAmt,
-		uint64(defaultGenesisTime.Unix()),
-		delegatorEndTime,
-		delegatorNodeID,
-		ids.GenerateTestShortID(),
-		[]*crypto.PrivateKeySECP256K1R{keys[0]},
-		keys[0].PublicKey().Address(), // change addr
-	)
-	assert.NoError(err)
-
-	service.vm.internalState.AddCurrentStaker(tx, 0)
-	service.vm.internalState.AddTx(tx, status.Committed)
-	err = service.vm.internalState.Commit()
-	assert.NoError(err)
-	err = service.vm.internalState.(*internalStateImpl).loadCurrentValidators()
-	assert.NoError(err)
-
-	// Make sure the delegator addr has the right stake (old stake + stakeAmt)
-	addr, _ := service.vm.FormatLocalAddress(keys[0].PublicKey().Address())
-	args.Addresses = []string{addr}
-	err = service.GetStake(nil, &args, &response)
-	assert.NoError(err)
-	assert.EqualValues(oldStake+stakeAmt, uint64(response.Staked))
-	assert.Len(response.Outputs, 2)
-	// Unmarshal into transferableoutputs
-	outputs := make([]avax.TransferableOutput, 2)
-	for i := range outputs {
-		outputBytes, err := formatting.Decode(args.Encoding, response.Outputs[i])
-		assert.NoError(err)
-		_, err = Codec.Unmarshal(outputBytes, &outputs[i])
-		assert.NoError(err)
-	}
-	// Make sure the stake amount is as expected
-	assert.EqualValues(stakeAmt+oldStake, outputs[0].Out.Amount()+outputs[1].Out.Amount())
-
-	oldStake = uint64(response.Staked)
-
-	// Make sure this works for pending stakers
-	// Add a pending staker
-	stakeAmt = service.vm.MinValidatorStake + 54321
-	pendingStakerNodeID := ids.GenerateTestShortID()
-	pendingStakerEndTime := uint64(defaultGenesisTime.Add(defaultMinStakingDuration).Unix())
-	tx, err = service.vm.newAddValidatorTx(
-		stakeAmt,
-		uint64(defaultGenesisTime.Unix()),
-		pendingStakerEndTime,
-		pendingStakerNodeID,
-		ids.GenerateTestShortID(),
-		0,
-		[]*crypto.PrivateKeySECP256K1R{keys[0]},
-		keys[0].PublicKey().Address(), // change addr
-	)
-	assert.NoError(err)
-
-	service.vm.internalState.AddPendingStaker(tx)
-	service.vm.internalState.AddTx(tx, status.Committed)
-	err = service.vm.internalState.Commit()
-	assert.NoError(err)
-	err = service.vm.internalState.(*internalStateImpl).loadPendingValidators()
-	assert.NoError(err)
-
-	// Make sure the delegator has the right stake (old stake + stakeAmt)
-	err = service.GetStake(nil, &args, &response)
-	assert.NoError(err)
-	assert.EqualValues(oldStake+stakeAmt, response.Staked)
-	assert.Len(response.Outputs, 3)
-	outputs = make([]avax.TransferableOutput, 3)
-	// Unmarshal
-	for i := range outputs {
-		outputBytes, err := formatting.Decode(args.Encoding, response.Outputs[i])
-		assert.NoError(err)
-		_, err = Codec.Unmarshal(outputBytes, &outputs[i])
-		assert.NoError(err)
-	}
-	// Make sure the stake amount is as expected
-	assert.EqualValues(stakeAmt+oldStake, outputs[0].Out.Amount()+outputs[1].Out.Amount()+outputs[2].Out.Amount())
-}
-
-// Test method GetCurrentValidators
-func TestGetCurrentValidators(t *testing.T) {
-	service := defaultService(t)
-	defaultAddress(t, service)
-	service.vm.ctx.Lock.Lock()
-	defer func() {
-		if err := service.vm.Shutdown(); err != nil {
-			t.Fatal(err)
-		}
-		service.vm.ctx.Lock.Unlock()
-	}()
-
-	genesis, _ := defaultGenesis()
-
-	// Call getValidators
-	args := GetCurrentValidatorsArgs{SubnetID: constants.PrimaryNetworkID}
-	response := GetCurrentValidatorsReply{}
-
-	err := service.GetCurrentValidators(nil, &args, &response)
-	switch {
-	case err != nil:
-		t.Fatal(err)
-	case len(response.Validators) != len(genesis.Validators):
-		t.Fatalf("should be %d validators but are %d", len(genesis.Validators), len(response.Validators))
-	}
-
-	for _, vdr := range genesis.Validators {
-		found := false
-		for i := 0; i < len(response.Validators) && !found; i++ {
-			gotVdr, ok := response.Validators[i].(APIPrimaryValidator)
-			switch {
-			case !ok:
-				t.Fatal("expected APIPrimaryValidator")
-			case gotVdr.NodeID != vdr.NodeID:
-			case gotVdr.EndTime != vdr.EndTime:
-				t.Fatalf("expected end time of %s to be %v but got %v",
-					vdr.NodeID,
-					vdr.EndTime,
-					gotVdr.EndTime,
-				)
-			case gotVdr.StartTime != vdr.StartTime:
-				t.Fatalf("expected start time of %s to be %v but got %v",
-					vdr.NodeID,
-					vdr.StartTime,
-					gotVdr.StartTime,
-				)
-			case gotVdr.Weight != vdr.Weight:
-				t.Fatalf("expected weight of %s to be %v but got %v",
-					vdr.NodeID,
-					vdr.Weight,
-					gotVdr.Weight,
-				)
-			default:
-				found = true
-			}
-		}
-		if !found {
-			t.Fatalf("expected validators to contain %s but didn't", vdr.NodeID)
-		}
-	}
-
-	// Add a delegator
-	stakeAmt := service.vm.MinDelegatorStake + 12345
-	validatorNodeID := keys[1].PublicKey().Address()
-	delegatorStartTime := uint64(defaultValidateStartTime.Unix())
-	delegatorEndTime := uint64(defaultValidateStartTime.Add(defaultMinStakingDuration).Unix())
-
-	tx, err := service.vm.newAddDelegatorTx(
-		stakeAmt,
-		delegatorStartTime,
-		delegatorEndTime,
-		validatorNodeID,
-		ids.GenerateTestShortID(),
-		[]*crypto.PrivateKeySECP256K1R{keys[0]},
-		keys[0].PublicKey().Address(), // change addr
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	service.vm.internalState.AddCurrentStaker(tx, 0)
-	service.vm.internalState.AddTx(tx, status.Committed)
-	err = service.vm.internalState.Commit()
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = service.vm.internalState.(*internalStateImpl).loadCurrentValidators()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Call getCurrentValidators
-	args = GetCurrentValidatorsArgs{SubnetID: constants.PrimaryNetworkID}
-	err = service.GetCurrentValidators(nil, &args, &response)
-	switch {
-	case err != nil:
-		t.Fatal(err)
-	case len(response.Validators) != len(genesis.Validators):
-		t.Fatalf("should be %d validators but are %d", len(genesis.Validators), len(response.Validators))
-	}
-
-	// Make sure the delegator is there
-	found := false
-	for i := 0; i < len(response.Validators) && !found; i++ {
-		vdr := response.Validators[i].(APIPrimaryValidator)
-		if vdr.NodeID != validatorNodeID.PrefixedString(constants.NodeIDPrefix) {
-			continue
-		}
-		found = true
-		if len(vdr.Delegators) != 1 {
-			t.Fatalf("%s should have 1 delegator", vdr.NodeID)
-		}
-		delegator := vdr.Delegators[0]
-		switch {
-		case delegator.NodeID != vdr.NodeID:
-			t.Fatal("wrong node ID")
-		case uint64(delegator.StartTime) != delegatorStartTime:
-			t.Fatal("wrong start time")
-		case uint64(delegator.EndTime) != delegatorEndTime:
-			t.Fatal("wrong end time")
-		case delegator.weight() != stakeAmt:
-			t.Fatalf("wrong weight")
-		}
-	}
-	if !found {
-		t.Fatalf("didnt find delegator")
-	}
-}
+// ? @jax i should not just blindly comment this out, it tests removing a stake as a whole including a delegation
 
 func TestGetTimestamp(t *testing.T) {
 	assert := assert.New(t)
