+++ vms/platformvm/tx_heap_by_end_time_test.go
@@ -39,7 +39,6 @@ func TestTxHeapStop(t *testing.T) {
 		uint64(defaultGenesisTime.Add(defaultMinStakingDuration).Unix()+1), // endTime
 		ids.ShortID{},                           // node ID
 		ids.ShortID{1, 2, 3, 4, 5, 6, 7},        // reward address
-		0,                                       // shares
 		[]*crypto.PrivateKeySECP256K1R{keys[0]}, // key
 		ids.ShortEmpty,                          // change addr
 	)
@@ -54,7 +53,6 @@ func TestTxHeapStop(t *testing.T) {
 		uint64(defaultGenesisTime.Add(defaultMinStakingDuration).Unix()+2), // endTime
 		ids.ShortID{1},                          // node ID
 		ids.ShortID{1, 2, 3, 4, 5, 6, 7},        // reward address
-		0,                                       // shares
 		[]*crypto.PrivateKeySECP256K1R{keys[0]}, // key
 		ids.ShortEmpty,                          // change addr
 	)
@@ -69,7 +67,6 @@ func TestTxHeapStop(t *testing.T) {
 		uint64(defaultGenesisTime.Add(defaultMinStakingDuration).Unix()+3), // endTime
 		ids.ShortID{},                           // node ID
 		ids.ShortID{1, 2, 3, 4, 5, 6, 7},        // reward address
-		0,                                       // shares
 		[]*crypto.PrivateKeySECP256K1R{keys[0]}, // key
 		ids.ShortEmpty,                          // change addr
 	)
