+++ wallet/chain/p/backend.go
@@ -65,8 +65,6 @@ func (b *backend) AcceptTx(ctx stdcontext.Context, tx *platformvm.Tx) error {
 	var baseTx *platformvm.BaseTx
 	txID := tx.ID()
 	switch utx := tx.UnsignedTx.(type) {
-	case *platformvm.UnsignedAddDelegatorTx:
-		baseTx = &utx.BaseTx
 	case *platformvm.UnsignedAddSubnetValidatorTx:
 		baseTx = &utx.BaseTx
 	case *platformvm.UnsignedAddValidatorTx:
