+++ wallet/chain/p/signer.go
@@ -79,8 +79,6 @@ func (s *signer) Sign(ctx stdcontext.Context, tx *platformvm.Tx) error {
 		return s.signAddValidatorTx(ctx, tx, utx)
 	case *platformvm.UnsignedAddSubnetValidatorTx:
 		return s.signAddSubnetValidatorTx(ctx, tx, utx)
-	case *platformvm.UnsignedAddDelegatorTx:
-		return s.signAddDelegatorTx(ctx, tx, utx)
 	case *platformvm.UnsignedCreateChainTx:
 		return s.signCreateChainTx(ctx, tx, utx)
 	case *platformvm.UnsignedCreateSubnetTx:
@@ -112,14 +110,6 @@ func (s *signer) signAddSubnetValidatorTx(ctx stdcontext.Context, tx *platformvm
 		return err
 	}
 	txSigners = append(txSigners, subnetAuthSigners)
-	return s.sign(tx, txSigners)
-}
-
-func (s *signer) signAddDelegatorTx(ctx stdcontext.Context, tx *platformvm.Tx, utx *platformvm.UnsignedAddDelegatorTx) error {
-	txSigners, err := s.getSigners(ctx, constants.PlatformChainID, utx.Ins)
-	if err != nil {
-		return err
-	}
 	return s.sign(tx, txSigners)
 }
